fix(ci): restore python coverage shard — split oversized operator hint tests
