+++ tests/unit/runtime/test_pr_monitor_operator_hints_part_007.py
@@ -1,0 +1,208 @@
+"""Regression tests for operator remonitor hints — preserved-head cleanup slice."""
+
+from __future__ import annotations
+
+from collections.abc import AsyncIterator
+from pathlib import Path
+
+import pytest
+from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
+
+from awf.common.commands import FakeCommandRunner
+from awf.common.github_client import RepoRef
+from awf.db.session import make_session_factory
+from awf.runtime.pr_monitor import (
+    _PROTECTED_BLOCK_PRESERVED_HEAD_STATE_KEY,
+    MonitorState,
+    OperatorHint,
+)
+from awf.runtime.pr_monitor_runner.comments import VerdictResult
+from awf.runtime.pr_monitor_runner.remote_ops import _GitPushResult
+from tests.postgres import postgres_test_engine
+from tests.unit.runtime._monitor_runner_fixtures import (
+    FakeAdapter,
+    RecordedSleep,
+    make_runner,
+    seed_monitoring_workspace,
+)
+
+
+@pytest.fixture
+async def factory() -> AsyncIterator[async_sessionmaker[AsyncSession]]:
+    async with postgres_test_engine() as engine:
+        yield make_session_factory(engine)
+
+
+async def _seed_active_grant(
+    factory: async_sessionmaker[AsyncSession],
+    workspace_id: str,
+    *,
+    path: str,
+    block_epoch: int = 0,
+) -> str:
+    from awf.common.ids import new_operator_grant_id
+    from awf.db.models import OperatorGrantAuditRecord
+
+    grant_id = new_operator_grant_id()
+    async with factory() as session:
+        session.add(
+            OperatorGrantAuditRecord(
+                id=grant_id,
+                workspace_id=workspace_id,
+                operator="op@example.com",
+                reason="approved the protected change",
+                normalized_path=path,
+                block_epoch=block_epoch,
+                approve_policy_downgrade=True,
+            )
+        )
+        await session.commit()
+    return grant_id
+
+
+@pytest.mark.unit
+async def test_operator_hint_resume_push_clears_preserved_head_marker(
+    factory: async_sessionmaker[AsyncSession],
+    tmp_path: Path,
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    """After a grant-only resume pushes the approved commit and is finalized, the
+    preserved-head marker MUST be dropped from monitor state. Leaving it would let
+    a later plain remonitor (no directive, no grant) whose old preserved commit is
+    still on the remote take the restart-recovery shortcut and silently skip the
+    CLI, ignoring the operator's new repair request (PRRT_kwDOSJAM6s6KE2BX)."""
+    workspace_id = await seed_monitoring_workspace(factory)
+    await _seed_active_grant(factory, workspace_id, path="pyproject.toml")
+    runner = make_runner(
+        factory=factory,
+        cmd=FakeCommandRunner(),
+        adapter=FakeAdapter(),
+        sleep_fn=RecordedSleep(),
+        worktrees_root=tmp_path,
+    )
+    hint = OperatorHint(
+        reason="approved the protected change",
+        operation_id="op_clears_marker",
+        requested_at="2026-06-17T00:00:00+00:00",
+        reason_code="OPERATOR_GUIDE",
+    )
+    state = MonitorState(pending_operator_hint=hint)
+    state.mark_addressed(_PROTECTED_BLOCK_PRESERVED_HEAD_STATE_KEY, "recorded-preserved-sha")
+
+    async def _no_preexisting_dirty(**_kwargs: object) -> None:
+        return None
+
+    async def _start_head_ok(**_kwargs: object) -> tuple[str, None]:
+        return ("recorded-preserved-sha", None)
+
+    async def _cli_must_not_run(**_kwargs: object) -> object:
+        pytest.fail("a grant-only resume must NOT invoke the CLI")
+
+    async def _no_block(**_kwargs: object) -> None:
+        return None
+
+    async def _not_on_remote(**_kwargs: object) -> bool:
+        return False
+
+    async def _pushed(**_kwargs: object) -> _GitPushResult:
+        return _GitPushResult(pushed=True, failed=False, returncode=0)
+
+    async def _head(*_args: object, **_kwargs: object) -> str:
+        return "pushed-sha"
+
+    monkeypatch.setattr(runner, "_pre_existing_dirty_repair_worktree_result", _no_preexisting_dirty)
+    monkeypatch.setattr(runner, "_repair_operation_start_head_result", _start_head_ok)
+    monkeypatch.setattr(runner, "_invoke_cli_for_verdict_result", _cli_must_not_run)
+    monkeypatch.setattr(runner, "_protected_scope_push_block", _no_block)
+    monkeypatch.setattr(runner, "_preserved_commit_already_on_remote", _not_on_remote)
+    monkeypatch.setattr(runner, "_validated_git_push_result", _pushed)
+    monkeypatch.setattr(runner, "_rev_parse_head", _head)
+
+    result = await runner._run_operator_hint_cycle(
+        workspace_id=workspace_id,
+        repo=RepoRef(owner="dimileeh", name="aira-web"),
+        pr_number=42,
+        pr_head_sha="abc1234567890def",
+        hint=hint,
+        state=state,
+        remote_branch=f"awf/{workspace_id}",
+        compose_project="proj",
+        compose_file=tmp_path / "compose.yml",
+    )
+
+    assert result.pushed is True
+    assert state.pending_operator_hint is None  # hint processed
+    # The marker is gone, so a subsequent plain remonitor will run the CLI.
+    assert _PROTECTED_BLOCK_PRESERVED_HEAD_STATE_KEY not in state.threads_addressed_ids
+
+
+@pytest.mark.unit
+async def test_operator_hint_no_directive_no_grant_no_marker_still_runs_cli(
+    factory: async_sessionmaker[AsyncSession],
+    tmp_path: Path,
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    """A plain remonitor hint (no directive, no grant, NO preserved-head marker)
+    must still invoke the CLI — the restart short-circuit only applies to a
+    protected-block grant resume, never to a normal remonitor."""
+    workspace_id = await seed_monitoring_workspace(factory)
+    runner = make_runner(
+        factory=factory,
+        cmd=FakeCommandRunner(),
+        adapter=FakeAdapter(),
+        sleep_fn=RecordedSleep(),
+        worktrees_root=tmp_path,
+    )
+    hint = OperatorHint(
+        reason="re-examine the PR",
+        operation_id="op_plain_remonitor",
+        requested_at="2026-06-17T00:00:00+00:00",
+    )
+    state = MonitorState(pending_operator_hint=hint)
+    cli_calls: list[dict[str, object]] = []
+
+    async def _no_preexisting_dirty(**_kwargs: object) -> None:
+        return None
+
+    async def _start_head_ok(**_kwargs: object) -> tuple[str, None]:
+        return ("abc1234567890def", None)
+
+    async def _fixed_verdict(**kwargs: object) -> VerdictResult:
+        cli_calls.append(kwargs)
+        return VerdictResult(verdict="fix_committed")
+
+    async def _no_block(**_kwargs: object) -> None:
+        return None
+
+    async def _not_on_remote(**_kwargs: object) -> bool:
+        return False
+
+    async def _pushed(**_kwargs: object) -> _GitPushResult:
+        return _GitPushResult(pushed=True, failed=False, returncode=0)
+
+    async def _head(*_args: object, **_kwargs: object) -> str:
+        return "pushed-sha"
+
+    monkeypatch.setattr(runner, "_pre_existing_dirty_repair_worktree_result", _no_preexisting_dirty)
+    monkeypatch.setattr(runner, "_repair_operation_start_head_result", _start_head_ok)
+    monkeypatch.setattr(runner, "_invoke_cli_for_verdict_result", _fixed_verdict)
+    monkeypatch.setattr(runner, "_protected_scope_push_block", _no_block)
+    monkeypatch.setattr(runner, "_preserved_commit_already_on_remote", _not_on_remote)
+    monkeypatch.setattr(runner, "_validated_git_push_result", _pushed)
+    monkeypatch.setattr(runner, "_rev_parse_head", _head)
+
+    result = await runner._run_operator_hint_cycle(
+        workspace_id=workspace_id,
+        repo=RepoRef(owner="dimileeh", name="aira-web"),
+        pr_number=42,
+        pr_head_sha="abc1234567890def",
+        hint=hint,
+        state=state,
+        remote_branch=f"awf/{workspace_id}",
+        compose_project="proj",
+        compose_file=tmp_path / "compose.yml",
+    )
+
+    assert result.pushed is True
+    assert cli_calls  # the CLI ran for the plain remonitor
+    assert state.pending_operator_hint is None
